+++ frontend/Foro/ForoUpload/main.js
