+++ frontend/Foro/ForoVisualizacion/index.html
@@ -5,7 +5,9 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Campus</title>
+
     <link rel="stylesheet" href="style.css">
+    <link rel="stylesheet" href="../../globalStyle.css">
 </head>
 <body>
     <div class="FVPaginaCompleta">
