+++ frontend/Foro/index.html
@@ -30,8 +30,8 @@
             <a href="../PaginaInicio/PaginaInicio.html" class="text" id="itemNav">Inicio</a>
             <a href="../ResumenesPrincipal/index.html" class="text" id="itemNav">Resumenes</a>
             <a class="text" id="itemNav"><span class="navResumen">Foros</span></a>
-            <a href="../Intercambios/Intercambios.html" class="text" id="itemNav">Intercambios</a>
-            <a href="../ObjetosPerdidos/ObjetosPerdidos.html" class="text" id="itemObjetos">Objetos Perdidos</a>
+            <a href="../Intercambios/index.html" class="text" id="itemNav">Intercambios</a>
+            <a href="../ObjetosPerdidos/index.html" class="text" id="itemObjetos">Objetos Perdidos</a>
         </div>
 
         <div class="boton" id="login">
