+++ frontend/Intercambios/IntercambiosVisualizacion/index.html
@@ -5,7 +5,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Campus</title>
-    <link rel="stylesheet" href="IntercambiosVisualizacion.css">
+    <link rel="stylesheet" href="styles.css">
 </head>
 <body>
     <div class="FotoBotones">
@@ -32,6 +32,6 @@
         </div>
     </div>
 
-    <script src="IntercambiosVisualizacion.js"></script>
+    <script src="main.js"></script>
 </body>
 </html>
