+++ frontend/PaginaInicio/index.html
@@ -3,7 +3,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Portal</title>
-    <link rel="stylesheet" href="PaginaInicio.css">
+    <link rel="stylesheet" href="styles.css">
 </head>
 <body>
     <header class="barrasuperiorbotonespi">
@@ -31,6 +31,6 @@
         <footer></footer>
     </article>
 
-    <script src="PaginaInicio/PaginaInicio.js"></script>
+    <script src="main.js"></script>
 </body>
 </html>
