+++ frontend/ResumenesPrincipal/index.html
@@ -28,11 +28,11 @@
         </div>
 
         <div id="navItems">
-            <a href="../PaginaInicio/PaginaInicio.html" class="text" id="itemNav">Inicio</a>
+            <a href="../PaginaInicio/index.html" class="text" id="itemNav">Inicio</a>
             <a class="text" id="itemNav"><span class="navResumen">Resumenes</span></a>
             <a href="../Foro/index.html" class="text" id="itemNav">Foros</a>
-            <a href="../Intercambios/Intercambios.html" class="text" id="itemNav">Intercambios</a>
-            <a href="../ObjetosPerdidos/ObjetosPerdidos.html" class="text" id="itemObjetos">Objetos Perdidos</a>
+            <a href="../Intercambios/index.html" class="text" id="itemNav">Intercambios</a>
+            <a href="../ObjetosPerdidos/index.html" class="text" id="itemObjetos">Objetos Perdidos</a>
         </div>
 
         <div class="boton" id="login">
