+++ frontend/ResumenesPrincipal/uploadResumenes/index.html
@@ -14,7 +14,7 @@
     <link href="https://fonts.googleapis.com/css2?family=Kodchasan&display=swap" rel="stylesheet">
 
     <link rel="stylesheet" href="style.css">
-    <link rel="stylesheet" href="../globalStyle.css">
+    <link rel="stylesheet" href="../../globalStyle.css">
 
     <link href="https://cdn.jsdelivr.net/npm/aos@2.3.4/dist/aos.css" rel="stylesheet">
     <script src="https://cdn.jsdelivr.net/npm/aos@2.3.4/dist/aos.js"></script>
@@ -22,16 +22,16 @@
 <body class="body">
     <nav class="navBar" id="nav">
         <div id="tituloYLogo">
-            <img id="logo" src="../img/Logo.svg" alt="logo">
+            <img id="logo" src="../../img/Logo.svg" alt="logo">
             <label for="PORTAL" id="tituloLogo">P<span id="ort">ORT</span>AL</label>
         </div>
 
         <div id="navItems">
-            <a href="../PaginaInicio/PaginaInicio.html" class="text" id="itemNav">Inicio</a>
+            <a href="../../PaginaInicio/PaginaInicio.html" class="text" id="itemNav">Inicio</a>
             <a class="text" id="itemNav"><span class="navResumen">Resumenes</span></a>
-            <a href="../Foro/Foro.html" class="text" id="itemNav">Foros</a>
-            <a href="../Intercambios/Intercambios.html" class="text" id="itemNav">Intercambios</a>
-            <a href="../ObjetosPerdidos/ObjetosPerdidos.html" class="text" id="itemObjetos">Objetos Perdidos</a>
+            <a href="../../Foro/index.html" class="text" id="itemNav">Foros</a>
+            <a href="../../Intercambios/Intercambios.html" class="text" id="itemNav">Intercambios</a>
+            <a href="../../ObjetosPerdidos/ObjetosPerdidos.html" class="text" id="itemObjetos">Objetos Perdidos</a>
         </div>
 
         <div class="boton" id="login">
