+++ frontend/ResumenesPrincipal/uploadResumenes/style.css
@@ -63,8 +63,8 @@
     border: 2px solid black;
     border-radius: 30px;
     outline: none; 
-    background-size: 10%;
-    background-image: url(../img/Arrow.svg);
+    background-size: 5%;
+    background-image: url(../../img/Arrow.svg);
     background-repeat: no-repeat;
     background-position: 92.5%;
 }
