+++ frontend/ResumenesPrincipal/visualizacionResumenes/main.js
@@ -1,16 +1,16 @@
+import { getQueryParams } from "../../controllers/detailsPageController.js";
 
-import { getQueryParams } from "../controllers/detailsPageController.js";
-console.log('Running preguntas visualizacion')
+console.log('Running resumen visualizacion')
 
-async function getPreguntasDetails() {
-    console.log('Getting preguntas info')
+async function getResumenesDetails() {
+    console.log('Getting resumen info')
     const token = localStorage.getItem('token')
 
     const id = await getQueryParams()
 
     try {
 
-        const response = await fetch(`http://localhost:3000/foros/open?id=${id}`, {
+        const response = await fetch(`http://localhost:3000/resumen/visualizar?id=${id}`, {
             method: 'GET',
             headers: {
                 'Authorization': `Bearer ${token}`,
@@ -19,11 +19,11 @@ async function getPreguntasDetails() {
     
         if (response.ok) {
 
-            console.log('Got the pregunta info')
+            console.log('Got the resumen info')
             const data = await response.json()
 
             console.log('Received data: ', data)
-            displayPregunta(data)
+            displayResumen(data)
         } 
 
     } catch (error) {
@@ -31,14 +31,14 @@ async function getPreguntasDetails() {
     }
 }
 
-
-async function displayPregunta(pregunta) {
+async function displayResumen(resumen) {
     const informacion = document.getElementById('informacion')
     const titulo = document.getElementById('titulo')
 
-    titulo.innerText = pregunta.titulo
-    informacion.innerText = pregunta.descripcion
+    titulo.innerText = resumen.titulo
+    informacion.innerText = resumen.descripcion
 
 }
 
-getPreguntasDetails()
+getResumenesDetails()
+
